modules/adapter/infrastructure/sqlalchemy/repository/kapt_repository.py
from typing import Callable, AsyncContextManager

from sqlalchemy.ext.asyncio import AsyncSession

from core.domain.kapt.interface.kapt_repository import KaptRepository
from modules.adapter.infrastructure.sqlalchemy.persistence.model.datalake.kapt_basic_info_model import (
    KaptBasicInfoModel,
)
from modules.adapter.infrastructure.sqlalchemy.repository import BaseAsyncRepository
from modules.adapter.infrastructure.utils.log_helper import logger_

logger = logger_.getLogger(__name__)


class AsyncKaptRepository(KaptRepository, BaseAsyncRepository):
    def __init__(
        self, session_factory: Callable[..., AsyncContextManager[AsyncSession]]
    ):
        super().__init__(session_factory=session_factory)

    async def find_by_id(self, house_id: int):
        async with self.session_factory() as session:
            kapt_basic_info = await session.get(KaptBasicInfoModel, house_id)

        if not kapt_basic_info:
            return None

        return kapt_basic_info.to_open_api_input_entity()

    # async def find_all
